pyproject.toml
[project]
name = "treequest"
version = "0.2.0"
description = "A Tree Search Library with Flexible API for LLM Inference-Time Scaling"
readme = "README.md"
authors = [
    { name = "Kou Misaki", email = "kou.misaki@sakana.ai" },
    { name = "Takuya Akiba", email = "takiba@sakana.ai" },
]
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0.0",
    "joblib>=1.5.2",
    "scipy>=1.0.0",
    "ulid-py>=1.1.0",
]

[project.optional-dependencies]
abmcts-m = [
    "jax>=0.5.2",
    "numpy>=1.26.4",
    "numpyro>=0.17.0",
    "pandas>=2.2.3",
    "pymc>=5.21.0",
    "xarray>=2025.1.2",
    "packaging>=23",
]

vis = [
    "Pillow>=10.0.0",
    "graphviz>=0.20.3",
    "jinja2>=3.1.0",
]

all = [
    "treequest[abmcts-m]",
    "treequest[vis]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "line-profiler>=4.2.0",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.4",
    "pytest-xdist>=3.6.1",
    "ruff>=0.9.9",
    "tqdm>=4.64.0",
    "types-tqdm>=4.64.0",
    "types-PyYAML>=6.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/treequest"]
include = [
    "src/treequest/vis/assets/colormaps.json",
    "src/treequest/vis/assets/d3.LICENSE.txt",
    "src/treequest/vis/assets/d3.v7.min.js",
    "src/treequest/vis/assets/d3_tree.html.jinja2",
]

[tool.hatch.build.targets.sdist]
include = [
    "src/treequest/vis/assets/colormaps.json",
    "src/treequest/vis/assets/d3.LICENSE.txt",
    "src/treequest/vis/assets/d3.v7.min.js",
    "src/treequest/vis/assets/d3_tree.html.jinja2",
]
